+++ packages/charts/react-charting/src/components/ScatterChart/ScatterChart.styles.ts
@@ -1,19 +1,9 @@
 import { IScatterChartStyleProps, IScatterChartStyles } from './ScatterChart.types';
+import { getMarkerLabelStyle, getTooltipStyle } from '../../utilities/index';
 
 export const getStyles = (props: IScatterChartStyleProps): IScatterChartStyles => {
   return {
-    tooltip: {
-      ...props.theme!.fonts.medium,
-      display: 'flex',
-      flexDirection: 'column',
-      padding: '8px',
-      position: 'absolute',
-      textAlign: 'center',
-      top: '0px',
-      background: props.theme!.semanticColors.bodyBackground,
-      borderRadius: '2px',
-      pointerEvents: 'none',
-      color: props.theme!.semanticColors.bodyText,
-    },
+    tooltip: getTooltipStyle(props.theme!),
+    markerLabel: getMarkerLabelStyle(props.theme!),
   };
 };