+++ packages/charts/react-charting/src/utilities/SVGTooltipText.tsx
@@ -3,7 +3,7 @@ import { ITooltipHost, ITooltipProps, Tooltip, TooltipDelay } from '@fluentui/re
 import { Async, KeyCodes, getId, portalContainsElement } from '../Utilities';
 import { ITheme } from '@fluentui/react';
 
-interface ISVGTooltipTextProps {
+export interface ISVGTooltipTextProps {
   /**
    * Number of milliseconds to delay closing the tooltip, so that the user has time to hover over
    * the tooltip and interact with it. Hovering over the tooltip will count as hovering over the
@@ -74,6 +74,11 @@ interface ISVGTooltipTextProps {
    * Prop to selectively display the tooltip background
    */
   showBackground?: boolean;
+
+  /**
+   * Prop to set classname
+   */
+  className?: string;
 }
 
 interface ISVGTooltipTextState {
@@ -85,7 +90,7 @@ interface ISVGTooltipTextState {
   textHeight?: number;
 }
 
-const PADDING = 4;
+const PADDING = 3;
 
 /**
  * Component to render an SVG text element with an optional tooltip.
@@ -139,9 +144,8 @@ export class SVGTooltipText
 
     const showTooltip =
       (!!this.props.isTooltipVisibleProp && this.state.isOverflowing && !!content) || (isTooltipVisible && !!content);
-    const backgroundColor = this.props.theme ? this.props.theme.semanticColors.bodyBackground : undefined;
     const rectX = (typeof textProps?.x === 'number' ? textProps.x : 0) - (textWidth ?? 0) / 2 - PADDING;
-    const rectY = (typeof textProps?.y === 'number' ? textProps.y : 0) - (textHeight ?? 0) / 2 - PADDING;
+    const rectY = (typeof textProps?.y === 'number' ? textProps.y : 0) - (textHeight ?? 0) / 2 - 2 * PADDING;
     return (
       <>
         {showBackground && (
@@ -149,9 +153,9 @@ export class SVGTooltipText
             x={rectX}
             y={rectY}
             width={(textWidth ?? 0) + 2 * PADDING}
-            height={(textHeight ?? 0) + 2 * PADDING}
-            fill={backgroundColor}
+            height={(textHeight ?? 0) + PADDING}
             transform={textProps?.transform}
+            className={this.props.className}
           />
         )}
         <text