+++ packages/charts/react-charts/library/src/components/ResponsiveContainer/index.ts
@@ -1,1 +1,3 @@
 export * from './ResponsiveContainer';
+export * from './ResponsiveContainer.types';
+export * from './withResponsiveContainer';