+++ packages/react-components/react-nav/library/src/components/NavSectionHeader/NavSectionHeader.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
+
 import { useNavSectionHeader_unstable } from './useNavSectionHeader';
 import { renderNavSectionHeader_unstable } from './renderNavSectionHeader';
 import { useNavSectionHeaderStyles_unstable } from './useNavSectionHeaderStyles.styles';
@@ -12,9 +14,8 @@ export const NavSectionHeader: ForwardRefComponent<NavSectionHeaderProps> = Reac
   const state = useNavSectionHeader_unstable(props, ref);
 
   useNavSectionHeaderStyles_unstable(state);
-  // TODO update types in packages/react-components/react-shared-contexts/src/CustomStyleHooksContext/CustomStyleHooksContext.ts
-  // https://github.com/microsoft/fluentui/blob/master/rfcs/react-components/convergence/custom-styling.md
-  // useCustomStyleHook_unstable('useNavSectionHeaderStyles_unstable')(state);
+  useCustomStyleHook_unstable('useNavSectionHeaderStyles_unstable')(state);
+
   return renderNavSectionHeader_unstable(state);
 });